src/secure-rag-agent/requirements.txt
dotenv
requests
azure-storage-blob
azure-ai-textanalytics
azure.keyvault
azure.identity
azure.search
azure.search.documents
six